servera/rest/requirements.txt
Flask
pymongo
flask-cors
jsonschema
jsonpath-ng
flask-socketio
eventlet
python-jose[cryptography]
requests